style: optimize copy button animation
