+++ .changeset/gorgeous-dancers-allow.md
@@ -1,0 +1,8 @@
+---
+"@rspress/plugin-preview": patch
+"@rspress/theme-default": patch
+"@rspress/runtime": patch
+"@rspress/core": patch
+---
+
+style: optimize code copy animation
