+++ packages/theme-default/src/layout/DocLayout/docComponents/code/index.module.scss
@@ -1,0 +1,47 @@
+.code-button-group {
+  position: absolute;
+  right: 20px;
+  top: 20px;
+  z-index: 10;
+  display: flex;
+  gap: 10px;
+  > button {
+    position: relative;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  svg {
+    opacity: inherit;
+    width: 20px;
+    height: 20px;
+    transition: all 0.3s ease;
+    color: var(--rp-c-text-3);
+    &:hover {
+      color: var(--rp-c-text-2);
+    }
+  }
+}
+
+.code-copy-button {
+  .icon-success {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%) scale(0);
+    opacity: 0;
+    color: #00d600;
+  }
+}
+
+.code-copied {
+  .icon-copy {
+    transform: scale(0.33);
+    opacity: 0;
+  }
+  .icon-success {
+    transform: translate(-50%, -50%) scale(1);
+    opacity: 1;
+    transition-delay: 0.075s;
+  }
+}
